requirements.txt
anthropic>=0.49.0
streamlit>=1.42.0
plotly>=5.24.0
python-dotenv>=1.0.0
